% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whole_adc.R
\name{adc_pr}
\alias{adc_pr}
\title{Transformation for precipitation values (whole method)}
\usage{
adc_pr(file_path_sim, file_path_obs)
}
\arguments{
\item{file_path_sim}{the path to the simulated precipitation values (already extracted)}

\item{file_path_obs}{the path to the observed precipitation values (.txt)}
}
\value{
The observed precipitation values with an added column for transformed values
}
\description{
Transformation for precipitation values (whole method)
}
\examples{
file_path_sim <- 'data/pr_sim.rda'
file_path_obs <- 'data/pr.rda'
trans <- adc_pr(file_path_sim, file_path_obs)
}
