% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvDeltaChange-package.R
\docType{package}
\name{AdvDeltaChange-package}
\alias{AdvDeltaChange}
\alias{AdvDeltaChange-package}
\title{AdvDeltaChange: Perform Advanced Delta Change on Data}
\description{
This package can extract data from rasters, derive the delta change parameters, transform observed series into climate change scenarios and visualize the results. It uses .txt files for observed data and .nc files for simulated data.
}
\keyword{internal}
