% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adc_par.R
\name{pr_a_nb}
\alias{pr_a_nb}
\title{Find the a parameter with no bias correction for precipitation}
\usage{
pr_a_nb(scen_q, ctrl_q, b)
}
\arguments{
\item{scen_q}{60 and 90 quantiles of the scenario period precipitation data}

\item{ctrl_q}{60 and 90 quantiles of the control period precipitation data}

\item{b}{Already calculated b parameter with pr_b_nb}
}
\value{
The a parameter value
}
\description{
Find the a parameter with no bias correction for precipitation
}
\examples{
load('data/pr_sim.rda')
pr_sim[, pr5 := 60*60*24*frollsum(pr, 5, align = 'c')]
pr_sim = pr_sim[!is.na(pr5)]
pr_ctrl = pr_sim[year(DTM) \%in\% 1981:2010]
pr_scen = pr_sim[year(DTM) \%in\% 2071:2100]
ctrl_q = quantile(pr_ctrl$pr5, c(.6,.9))
scen_q = quantile(pr_scen$pr5, c(.6,.9))
pr_b <- pr_b_nb(scen_q, ctrl_q)
pr_a <- pr_a_nb(scen_q, ctrl_q, pr_b)
}
