% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adc_par.R
\name{pr_b}
\alias{pr_b}
\title{Find the b parameter with bias correction for precipitation}
\usage{
pr_b(scen_q, ctrl_q, obs_q)
}
\arguments{
\item{scen_q}{60 and 90 quantiles of the scenario period precipitation data}

\item{ctrl_q}{60 and 90 quantiles of the control period precipitation data}

\item{obs_q}{60 and 90 quantiles of the observed precipitation data in the control period timeline}
}
\value{
The b parameter value
}
\description{
Find the b parameter with bias correction for precipitation
}
\examples{
load('data/pr_sim.rda')
pr_sim[, pr5 := 60*60*24*frollsum(pr, 5, align = 'c')]
pr_sim = pr_sim[!is.na(pr5)]
pr_ctrl = pr_sim[year(DTM) \%in\% 1981:2010]
pr_scen = pr_sim[year(DTM) \%in\% 2071:2100]
ctrl_q = quantile(pr_ctrl$pr5, c(.6,.9))
scen_q = quantile(pr_scen$pr5, c(.6,.9))
load('data/pr.rda')
pr[, pr_o5 := frollsum(pr_o, 5)]
pr = pr[year(DTM) \%in\% 1981:2010]
obs_q = quantile(pr$pr_o5, c(.6,.9))
b <- pr_b(scen_q, ctrl_q, obs_q)
}
