% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tas_transf.R
\name{pr_plot}
\alias{pr_plot}
\title{Visualization for precipitation values}
\usage{
pr_plot(pr)
}
\arguments{
\item{pr}{PR with transformed values (obtained with adc_pr())}
}
\value{
A plot of the transformed values
}
\description{
Visualization for precipitation values
}
\examples{
file_path_sim <- 'data/pr_sim.rda'
file_path_obs <- 'data/pr.rda'
pr_trans <- adc_pr(file_path_sim, file_path_obs)
pr_plot(pr_trans)
}
