% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tas_transf.R
\name{tas_plot}
\alias{tas_plot}
\title{Visualization for temperature values}
\usage{
tas_plot(tas)
}
\arguments{
\item{tas}{TAS with transformed values (obtained with adc_tas())}
}
\value{
A plot of the transformed values
}
\description{
Visualization for temperature values
}
\examples{
file_path_sim <- 'data/tas_sim.rda'
file_path_obs <- 'data/tas.rda'
tas_trans <- adc_tas(file_path_sim, file_path_obs)
tas_plot(tas_trans)
}
