\name{tas_sim}
\alias{tas_sim}
\docType{data}
\title{
A data table containing extracted values of temperature
}
\description{
Extracted from the .nc temperature file, the dataset shows the simulated values as a data table. Product of the rastr_extract function.
}
\usage{data("tas_sim")}
\format{
  A data table with 55115 observations on the following 2 variables.
  \describe{
    \item{\code{tas}}{values of simulated temperature}
    \item{\code{DTM}}{Dates for the values YYYY-MM-DD}
\examples{
data(tas_sim)
}
\keyword{datasets}
