% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{tas_transf}
\alias{tas_transf}
\title{Transformation for temperature values}
\usage{
tas_transf(tas, tas_scen, tas_ctrl)
}
\arguments{
\item{tas}{Observed temperature values (must contain a column named 'tas_o' for the values)}

\item{tas_scen}{Simulated scenario temperature values (the values column must be named 'tas')}

\item{tas_ctrl}{Simulated control temperature values (the values column must be named 'tas')}
}
\value{
The observed temperature values with an added column for transformed values
}
\description{
Transformation for temperature values
}
\examples{
load('data/tas_sim.rda')
tas_sim[, tas := tas-273.15]
tas_ctrl = tas_sim[year(DTM) \%in\% 1981:2010]
tas_scen = tas_sim[year(DTM) \%in\% 2071:2100]
load('data/tas.rda')
tas <- tas_transf(tas, tas_scen, tas_ctrl)
}
