% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whole_adc.R
\name{adc_pr}
\alias{adc_pr}
\title{Transformation for precipitation values (whole method)}
\usage{
adc_pr(data_sim, data_obs)
}
\arguments{
\item{data_sim}{the data with the simulated precipitation values (already extracted)}

\item{data_obs}{the data with the observed precipitation values}
}
\value{
The observed precipitation values with an added column for transformed values
}
\description{
Transformation for precipitation values (whole method)
}
\examples{
data_sim <- load('data/pr_sim.rda')
data_obs <- load('data/pr.rda')
trans <- adc_pr(data_sim, data_obs)
}
