% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whole_adc.R
\name{adc_tas}
\alias{adc_tas}
\title{Transformation for temperature values (whole method)}
\usage{
adc_tas(file_path_sim, file_path_obs)
}
\arguments{
\item{file_path_sim}{the path to the simulated temperature values (already extracted)}

\item{file_path_obs}{the path to the observed temperature values (.txt)}
}
\value{
The observed temperature values with an added column for transformed values
}
\description{
Transformation for temperature values (whole method)
}
\examples{
file_path_sim <- 'data/tas_sim.rda'
file_path_obs <- 'data/tas.rda'
tas_trans <- adc_tas(file_path_sim, file_path_obs)
}
